+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add support for setting the app CR name.
+- Set `app.kubernetes.io/name` label for app CRs.
+
 ## [0.9.0] - 2020-12-15
 
 ### Added